refactor out declarations code generation
To generate the solutions, one approach is to write our generated puzzle
code to a file and just run it. To do that, it'll be convenient to be
able to get that code as a string independently of the other puzzle
generation.

This commit does that refactoring. Generating the solution will be
later.