actually do non-default ctor obj creation
Before, the isEmpty() in ChoiceDeck looked at the shuffled deck -- which
is only populated after a call to draw(). But what we really are asking
when we call isEmpty is, are there any cards at all in this deck? So use
cards for that check.

This also fixes an infinite loop bug -- if the classes passed didn't
have any classes with a non-default constructor, you'd loop forever
drawing from the choice deck trying to find one.

Now we filter the list of classes and return Optional.empty() right away
if there's none with a non-default constructor.